Agent userdata tests fixups for 4.1.1.1-build (#551)
* Userdata agent test: add REST_HOST

* Userdata agent tests: add MANAGER_FILE_SERVER_ROOT_KEY

It is actually the path where the script is written, on the local
machine

* Userdata agent tests: create the install script dir

It lives under cloudify_agent/ in the fileserver root, so we need
to create cloudify_agent. Need to consider if maybe the cloudify-agent
function might need to do this, if the cloudify_agent subdir can also
not exist on the manager in some scenarios

* Userdata agent tests: .pop() with default when cleaning env

For some reason i get keyerrors on this? Is this just because of
the iterating over the dict while mutating it? (which is what
.iterkeys does here). Also changed it to not do that, with list() to
copy the keys