CY-2393 Remove agent userdata tests
These tests didn't make sense as they weren't something we'd support in the wild
(because of the complicated set of steps required to reach it followed by using
the 2.x Openstack plugin but not the 3.x one).
A jira will be raised to create some sensible agent tests for this.