RD-858 test_inplace_restore: wait_for_manager after snap-restore (#1232)
* RD-858 test_inplace_restore: wait_for_manager after snap-restore

After restoring the snapshot, we wait for the manager to come back
up, because the snapshot-restore includes a reboot.

Only THEN we wait for agents. Why would we wait for agents before
the manager was up?

The order of operations is:
1. snap-restore
2. reboot
3. wait for manager to come back up
4. wait for agents to reconnect

Before this change, 3 and 4 were swapped

* RD-859 reduce the wait time for agents

60 seconds is too long. The log says that 30 seconds should be
enough. Let's give it 35 to allow for some slip.

Also, really, most likely the agents will be able to reconnect during
the wait-for-manager phase anyway, just when rabbitmq is up.