Remove agents dir on teardown test
The agent itself gets removed, leaving the dir empty, but then the dir isn't cleaned up.
As there could technically be multiple agents this is maybe not a bug on the agent side.
However, as we know there won't be multiple agents, we should be able to remove it in the
test, and fail if we can't.