Make cleanup script cope with untagged resources
The VPC MUST be tagged, but we will now delete all child resources
of any correctly tagged VPC when we try to.

The only resource we can now leak is EIPs because they don't belong
to a VPC, so if they're not tagged... sadness is a thing.

Estimate age will also return hours above 24 now.