RND-901 Download new rest cert: re-establish the http session
Close the session after fetching the cert, so that next request
re-loads the cert, instead of using the old cert and throwing an error.

Alternatively, we could just send a dummy request, that would fail
with a SSLError - that also causes a reload, so the next request would
pass. But that would be silly :)